+++ src/nupic/utils/SdrMetrics.hpp
@@ -23,18 +23,14 @@
 #define SDR_METRICS_HPP
 
 #include <vector>
-#include <numeric>
 #include <nupic/types/Sdr.hpp>
 #include <nupic/types/Types.hpp>
-#include <nupic/types/Serializable.hpp>
-#include <nupic/utils/Random.hpp>
-#include <cmath> // std::log2 isnan
-#include <regex>
 
 using namespace std; // TODO Remove this!
 
 namespace nupic {
 namespace sdr {
+
 
 /**
  * Helper for SDR metrics trackers, including: Sparsity,
@@ -43,13 +39,30 @@ namespace sdr {
  * Subclasses must override method "callback".
  */
 class MetricsHelper_ {
-protected:
-    UInt  period_;
-    UInt  samples_;
+public:
+    const UInt         &period     = period_;
+    const UInt         &samples    = samples_;
+    const vector<UInt> &dimensions = dimensions_;
+
+    /**
+     * Add an SDR datum to this Metric.  This method can only be called if the
+     * Metric was constructed with dimensions and NOT an SDR.
+     *
+     * The given SDR's dimensions must be the same as this Metric's dimensions.
+     */
+    void addData(SDR &data);
+
+    virtual ~MetricsHelper_();
+
+private:
     vector<UInt> dimensions_;
-    SDR*  dataSource_;
-    UInt  callback_handle_;
-    UInt  destroyCallback_handle_;
+    SDR* dataSource_;
+    UInt callback_handle_;
+    UInt destroyCallback_handle_;
+
+protected:
+    UInt period_;
+    UInt samples_;
 
     /**
      * @param dimensions of SDR.  Add data to this metric by calling method
@@ -57,16 +70,7 @@ protected:
      *
      * @param period Time scale for exponential moving average.
      */
-    MetricsHelper_( const vector<UInt> dimensions, UInt period ) {
-        NTA_CHECK( period > 0u );
-        NTA_CHECK( dimensions.size() > 0 );
-        dimensions_ = dimensions,
-        period_     = period;
-        samples_    = 0u;
-        dataSource_ = nullptr;
-        callback_handle_        = -1;
-        destroyCallback_handle_ = -1;
-    }
+    MetricsHelper_( const vector<UInt> dimensions, UInt period );
 
     /**
      * @param dataSource SDR to track.  Add data to the metric by assigning to
@@ -75,28 +79,13 @@ protected:
      *
      * @param period Time scale for exponential moving average.
      */
-    MetricsHelper_( SDR &dataSource, UInt period )
-        : MetricsHelper_(dataSource.dimensions, period)
-    {
-        dataSource_ = &dataSource;
-        callback_handle_ = dataSource_->addCallback( [&](){
-            callback( *dataSource_, 1.0f / std::min( period_, (UInt) ++samples_ ));
-        });
-        destroyCallback_handle_ = dataSource_->addDestroyCallback( [&](){
-            deconstruct();
-        });
-    }
+    MetricsHelper_( SDR &dataSource, UInt period );
 
-    void deconstruct() {
-        if( dataSource_ != nullptr ) {
-            dataSource_->removeCallback( callback_handle_ );
-            dataSource_->removeDestroyCallback( destroyCallback_handle_ );
-            dataSource_ = nullptr;
-        }
-    }
+    void deconstruct();
 
     /**
      * Add another datum to the metric.
+     *      Subclasses MUST override this method!
      *
      * @param dataSource SDR to add to the metric.
      *
@@ -106,29 +95,8 @@ protected:
      * sample.
      */
     virtual void callback( SDR &dataSource, Real alpha ) = 0;
-
-public:
-    const UInt         &samples    = samples_;
-    const UInt         &period     = period_;
-    const vector<UInt> &dimensions = dimensions_;
-
-    /**
-     * Add an SDR datum to this Metric.  This method can only be called if the
-     * Metric was constructed with dimensions and NOT an SDR.
-     *
-     * The given SDR's dimensions must be the same as this Metric's dimensions.
-     */
-    void addData(SDR &data) {
-        NTA_CHECK( dataSource_ == nullptr )
-            << "Method addData can only be called if this metric was NOT initialize with an SDR!";
-        NTA_CHECK( dimensions_ == data.dimensions );
-        callback( data, 1.0f / std::min( period_, (UInt) ++samples_ ));
-    }
-
-    virtual ~MetricsHelper_() {
-        deconstruct();
-    }
 };
+
 
 /**
  * Sparsity class
@@ -151,33 +119,6 @@ public:
  *      cout << B  -> Sparsity Min/Mean/Std/Max 0.01 / 0.0700033 / 0.0588751 / 0.15
  */
 class Sparsity : public MetricsHelper_ {
-private:
-    Real min_;
-    Real max_;
-    Real mean_;
-    Real variance_;
-    Real sparsity_;
-
-    void initialize() {
-        sparsity_   =  1234.567f;
-        min_        =  1234.567f;
-        max_        = -1234.567f;
-        mean_       =  1234.567f;
-        variance_   =  1234.567f;
-    }
-
-    void callback(SDR &dataSource, Real alpha) override {
-        sparsity_ = dataSource.getSparsity();
-        min_ = std::min( min_, sparsity_ );
-        max_ = std::max( max_, sparsity_ );
-        // http://people.ds.cam.ac.uk/fanf2/hermes/doc/antiforgery/stats.pdf
-        // See section 9.
-        const Real diff      = sparsity_ - mean_;
-        const Real incr      = alpha * diff;
-                   mean_    += incr;
-                   variance_ = (1.0f - alpha) * (variance_ + diff * incr);
-    }
-
 public:
     /**
      * @param dataSource SDR to track.  Add data to this sparsity metric by
@@ -185,9 +126,7 @@ public:
      *
      * @param period Time scale for exponential moving average.
      */
-    Sparsity( SDR &dataSource, UInt period )
-        : MetricsHelper_( dataSource, period )
-        { initialize(); }
+    Sparsity( SDR &dataSource, UInt period );
 
     /**
      * @param dimensions of SDR.  Add data to this sparsity metric by calling
@@ -195,23 +134,29 @@ public:
      *
      * @param period Time scale for exponential moving average.
      */
-    Sparsity( const vector<UInt> dimensions, UInt period )
-        : MetricsHelper_( dimensions, period )
-        { initialize(); }
+    Sparsity( const vector<UInt> dimensions, UInt period );
 
     const Real &sparsity = sparsity_;
-    Real min() const { return min_; }
-    Real max() const { return max_; }
-    Real mean() const { return mean_; }
-    Real std() const { return std::sqrt( variance_ ); }
 
-    friend std::ostream& operator<<(std::ostream& stream, const Sparsity &S)
-    {
-        return stream << "Sparsity Min/Mean/Std/Max "
-            << S.min() << " / " << S.mean() << " / "
-            << S.std() << " / " << S.max() << endl;
-    }
+    Real min() const;
+    Real max() const;
+    Real mean() const;
+    Real std() const;
+
+    friend std::ostream& operator<< ( std::ostream &, const Sparsity & );
+
+private:
+    Real sparsity_;
+    Real min_;
+    Real max_;
+    Real mean_;
+    Real variance_;
+
+    void initialize();
+
+    void callback(SDR &dataSource, Real alpha) override;
 };
+
 
 /**
  * ActivationFrequency class
@@ -240,24 +185,6 @@ public:
  *                     Entropy 0.918296
  */
 class ActivationFrequency : public MetricsHelper_ {
-private:
-    vector<Real> activationFrequency_;
-
-    void initialize(UInt size) {
-        activationFrequency_.assign( size, 1234.567f );
-    }
-
-    void callback(SDR &dataSource, Real alpha) override
-    {
-        const auto decay = 1.0f - alpha;
-        for(auto &value : activationFrequency_)
-            value *= decay;
-
-        const auto &sparse = dataSource.getSparse();
-        for(const auto &idx : sparse)
-            activationFrequency_[idx] += alpha;
-    }
-
 public:
     /**
      * @param dataSource SDR to track.  Add data to this ActivationFrequency
@@ -265,9 +192,7 @@ public:
      *
      * @param period Time scale for exponential moving average.
      */
-    ActivationFrequency( SDR &dataSource, UInt period )
-        : MetricsHelper_( dataSource, period )
-        { initialize( dataSource.size ); }
+    ActivationFrequency( SDR &dataSource, UInt period );
 
     /**
      * @param dimensions of SDR.  Add data to this ActivationFrequency
@@ -276,55 +201,16 @@ public:
      *
      * @param period Time scale for exponential moving average.
      */
-    ActivationFrequency( const vector<UInt> dimensions, UInt period )
-        : MetricsHelper_( dimensions, period )
-        {
-            UInt size = 1;
-            for(const auto &dim : dimensions)
-                size *= dim;
-            initialize( size );
-        }
+    ActivationFrequency( const vector<UInt> dimensions, UInt period );
 
     const vector<Real> &activationFrequency = activationFrequency_;
 
-    Real min() const {
-        return *std::min_element(activationFrequency_.begin(),
-                                 activationFrequency_.end());
-    }
+    Real min() const;
+    Real max() const;
+    Real mean() const;
+    Real std() const;
 
-    Real max() const {
-        return *std::max_element(activationFrequency_.begin(),
-                                 activationFrequency_.end());
-    }
-
-    Real mean() const  {
-        const auto sum = std::accumulate( activationFrequency_.begin(),
-                                          activationFrequency_.end(),
-                                          0.0f);
-        return (Real) sum / activationFrequency_.size();
-    }
-
-    Real std() const {
-        const auto mean_ = mean();
-        auto sum_squares = 0.0f;
-        for(const auto &frequency : activationFrequency) {
-            const auto displacement = frequency - mean_;
-            sum_squares += displacement * displacement;
-        }
-        const auto variance = sum_squares / activationFrequency.size();
-
-        return std::sqrt( variance );
-    }
-
-    static Real binary_entropy_(const vector<Real> &frequencies) {
-        Real accumulator = 0.0f;
-        for(const auto &p  : frequencies) {
-            const auto  p_ = 1.0f - p;
-            const auto  e  = -p * std::log2( p ) - p_ * std::log2( p_ );
-            accumulator   += isnan(e) ? 0.0f : e;
-        }
-        return accumulator / frequencies.size();
-    }
+    static Real binary_entropy_(const vector<Real> &frequencies);
 
     /**
      * Binary entropy is a measurement of information.  It measures how well the
@@ -337,21 +223,16 @@ public:
      *
      * @returns Binary entropy of SDR, scaled to range [0, 1].
      */
-    Real entropy() const {
-        const auto max_extropy = binary_entropy_({ mean() });
-        if( max_extropy == 0.0f )
-            return 0.0f;
-        return binary_entropy_( activationFrequency ) / max_extropy;
-    }
+    Real entropy() const;
 
-    friend std::ostream& operator<< (std::ostream& stream,
-                                     const ActivationFrequency &F)
-    {
-        stream << "Activation Frequency Min/Mean/Std/Max "
-            << F.min() << " / " << F.mean() << " / "
-            << F.std() << " / " << F.max() << endl;
-        return stream << "Entropy " << F.entropy() << endl;
-    }
+    friend std::ostream& operator<< (std::ostream &, const ActivationFrequency &);
+
+private:
+    vector<Real> activationFrequency_;
+
+    void initialize(UInt size);
+
+    void callback(SDR &dataSource, Real alpha) override;
 };
 
 
@@ -381,6 +262,34 @@ public:
  *      cout << B   -> Overlap Min/Mean/Std/Max 0.05 / 0.260016 / 0.16389 / 0.45
  */
 class Overlap : public MetricsHelper_ {
+public:
+    /**
+     * @param dataSource SDR to track.  Add data to this Overlap instance
+     * by assigning to this SDR.
+     *
+     * @param period Time scale for exponential moving average.
+     */
+    Overlap( SDR &dataSource, UInt period );
+
+    /**
+     * @param dimensions of SDR.  Add data to this Overlap instance
+     * by calling method addData(SDR&) with an SDR which has these dimensions.
+     *
+     * @param period Time scale for exponential moving average.
+     */
+    Overlap( const vector<UInt> dimensions, UInt period );
+
+    void reset();
+
+    const Real &overlap = overlap_;
+
+    Real min() const;
+    Real max() const;
+    Real mean() const;
+    Real std() const;
+
+    friend std::ostream& operator<< ( std::ostream &, const Overlap & );
+
 private:
     SDR  previous_;
     bool previousValid_;
@@ -390,76 +299,9 @@ private:
     Real mean_;
     Real variance_;
 
-    void initialize() {
-        overlap_    =  1234.567f;
-        min_        =  1234.567f;
-        max_        = -1234.567f;
-        mean_       =  1234.567f;
-        variance_   =  1234.567f;
-        reset();
-    }
+    void initialize();
 
-    void callback(SDR &dataSource, Real alpha) override {
-        if( not previousValid_ ) {
-            previous_.setSDR( dataSource );
-            previousValid_ = true;
-            // It takes two data samples to compute overlap so decrement the
-            // samples counter & return & wait for the next sample.
-            samples_ -= 1;
-            return;
-        }
-        const auto nbits = std::max( previous_.getSum(), dataSource.getSum() );
-        const auto rawOverlap = previous_.getOverlap( dataSource );
-        overlap_ = (nbits == 0u) ? 1.0f : (Real) rawOverlap / nbits;
-        min_     = std::min( min_, overlap_ );
-        max_     = std::max( max_, overlap_ );
-        // http://people.ds.cam.ac.uk/fanf2/hermes/doc/antiforgery/stats.pdf
-        // See section 9.
-        const Real diff      = overlap_ - mean_;
-        const Real incr      = alpha * diff;
-                   mean_    += incr;
-                   variance_ = (1.0f - alpha) * (variance_ + diff * incr);
-        previous_.setSDR( dataSource );
-    }
-
-public:
-    /**
-     * @param dataSource SDR to track.  Add data to this Overlap instance
-     * by assigning to this SDR.
-     *
-     * @param period Time scale for exponential moving average.
-     */
-    Overlap( SDR &dataSource, UInt period )
-        : MetricsHelper_( dataSource, period ),
-          previous_( dataSource.dimensions )
-        { initialize(); }
-
-    /**
-     * @param dimensions of SDR.  Add data to this Overlap instance
-     * by calling method addData(SDR&) with an SDR which has these dimensions.
-     *
-     * @param period Time scale for exponential moving average.
-     */
-    Overlap( const vector<UInt> dimensions, UInt period )
-        : MetricsHelper_( dimensions, period ),
-          previous_( dimensions )
-        { initialize(); }
-
-    void reset()
-        { previousValid_ = false; }
-
-    const Real &overlap = overlap_;
-    Real min() const { return min_; }
-    Real max() const { return max_; }
-    Real mean() const { return mean_; }
-    Real std() const { return std::sqrt( variance_ ); }
-
-    friend std::ostream& operator<<(std::ostream& stream, const Overlap &V)
-    {
-        return stream << "Overlap Min/Mean/Std/Max "
-            << V.min() << " / " << V.mean() << " / "
-            << V.std() << " / " << V.max() << endl;
-    }
+    void callback(SDR &dataSource, Real alpha) override;
 };
 
 /**
@@ -494,12 +336,6 @@ public:
  */
 // TODO: Add flags to enable/disable which metrics this uses?
 class Metrics {
-private:
-    vector<UInt>        dimensions_;
-    Sparsity            sparsity_;
-    ActivationFrequency activationFrequency_;
-    Overlap             overlap_;
-
 public:
     /**
      * @param dataSource SDR to track.  Add data to this Metrics instance
@@ -507,12 +343,7 @@ public:
      *
      * @param period Time scale for exponential moving average.
      */
-    Metrics( SDR &dataSource, UInt period )
-        : dimensions_( dataSource.dimensions ),
-          sparsity_(            dataSource, period ),
-          activationFrequency_( dataSource, period ),
-          overlap_(             dataSource, period )
-          {};
+    Metrics( SDR &dataSource, UInt period );
 
     /**
      * @param dimensions of SDR.  Add data to this Metrics instance
@@ -520,15 +351,9 @@ public:
      *
      * @param period Time scale for exponential moving average.
      */
-    Metrics( const vector<UInt> dimensions, UInt period )
-        : dimensions_( dimensions ),
-          sparsity_(            dimensions, period ),
-          activationFrequency_( dimensions, period ),
-          overlap_(             dimensions, period )
-          {};
+    Metrics( const vector<UInt> dimensions, UInt period );
 
-    void reset()
-        { overlap_.reset(); }
+    void reset();
 
     const vector<UInt>        &dimensions          = dimensions_;
     const Sparsity            &sparsity            = sparsity_;
@@ -541,34 +366,15 @@ public:
      *
      * The given SDR's dimensions must be the same as this Metric's dimensions.
      */
-    void addData(SDR &data) {
-        sparsity_.addData( data );
-        activationFrequency_.addData( data );
-        overlap_.addData( data );
-    }
+    void addData(SDR &data);
 
-    friend std::ostream& operator<<(std::ostream& stream, const Metrics &M)
-    {
-        // Introduction line:  "SDR ( dimensions )"
-        stream << "SDR( ";
-        for(const auto &dim : M.dimensions_)
-            stream << dim << " ";
-        stream << ")" << endl;
+    friend std::ostream& operator<<(std::ostream& stream, const Metrics &M);
 
-        // Print data to temporary area for formatting.
-        stringstream data_stream;
-        data_stream << M.sparsity;
-        data_stream << M.activationFrequency;
-        data_stream << M.overlap;
-        string data = data_stream.str();
-
-        // Indent all of the data text (4 spaces).  Append indent to newlines.
-        data = regex_replace( data, regex("\n"), "\n\r    " );
-        // Strip trailing whitespace
-        data = regex_replace( data, regex("\\s+$"), "" );
-        // Insert first indent, append trailing newline.
-        return stream << "    " << data << endl;
-    }
+private:
+    vector<UInt>        dimensions_;
+    Sparsity            sparsity_;
+    ActivationFrequency activationFrequency_;
+    Overlap             overlap_;
 };
 
 } // end namespace sdr
